requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.2
Pillow>=10.0
python-decouple>=3.8
gunicorn>=22.0
psycopg2-binary>=2.9  # If using PostgreSQL in production